wacom: use long for data in props
as using integers would cause some garbage to be passed to
the Wacom driver on 64 bits systems where the size of int
is less than the size of long, confusing the Wacom driver.

https://bugzilla.gnome.org/show_bug.cgi?id=690588

+++ plugins/wacom/gsd-wacom-manager.c
@@ -625,7 +625,7 @@ reset_touch_buttons (XDevice               *xdev,
 	for (i = 0; buttons[i].button != NULL; i++)
 	{
 		char *propname;
-		int action[2]; /* press + release */
+		glong action[2]; /* press + release */
 		Atom prop;
 		int mapped_button = buttons[i].num;
